CHORE: Post 도메인 분리
