+++ src/main/java/com/flab/posttoy/domain/Post.java
@@ -3,16 +3,14 @@ package com.flab.posttoy.domain;
 import lombok.Builder;
 import lombok.Getter;
 
+
 // 불변객체를 유지하기 위해 setter 대신 builder
 @Getter
 @Builder
 public class Post {
-    private Long id;
+    @Builder.Default
+    private Long id = Long.valueOf(0L);
     private Long userId;
     private String title;
     private String content;
-
-    public void setId(Long id) {
-        this.id = id;
-    }
 }
